% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oss.s3-package.R
\docType{package}
\name{oss.s3-package}
\alias{oss.s3-package}
\alias{oss.s3}
\title{oss.s3-package}
\description{
Aliyun OSS Client Package
}
\details{
A simple client package for the Alibaba Cloud (Aliyun) Object Storage Service (OSS) REST API.
}
\author{
Ruiyang Gan <ganruiyang@metrodata.cn>
}
\keyword{package}
